requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via starlette
argon2-cffi==25.1.0
    # via pwdlib
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
cffi==2.0.0
    # via argon2-cffi-bindings
click==8.3.1
    # via uvicorn
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via pydantic
exceptiongroup==1.3.1
    # via anyio
fastapi==0.135.1
    # via -r requirements.in
greenlet==3.3.2
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   email-validator
mangum==0.21.0
    # via -r requirements.in
pwdlib[argon2]==0.3.0
    # via -r requirements.in
pycparser==3.0
    # via cffi
pydantic[email]==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.41.5
    # via pydantic
pyjwt==2.11.0
    # via -r requirements.in
python-dotenv==1.2.2
    # via -r requirements.in
python-multipart==0.0.22
    # via -r requirements.in
sqlalchemy==2.0.48
    # via -r requirements.in
starlette==0.52.1
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   mangum
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
uvicorn==0.41.0
    # via -r requirements.in
